requirements.txt
Babel==1.3
Django==1.8.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Pygments==2.0.2
Sphinx==1.3.1
alabaster==0.7.4
argparse==1.2.1
defusedxml==0.4.1
django-allauth==0.19.1
django-crispy-forms==1.4.0
django-registration-redux==1.2
django-rest-auth==0.4.0
django-rest-swagger==0.2.9
djangorestframework==3.1.2
docutils==0.12
httpie==0.9.2
oauthlib==0.7.2
python-openid==2.2.5
python3-openid==3.0.5
pytz==2015.4
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
snowballstemmer==1.2.0
sphinx-rtd-theme==0.1.8
